UI changes and code comments

+++ src/components/DataMethods.tsx
@@ -2,27 +2,6 @@ import { apiURL } from "../data/API_URL";
 import { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
 
-export const fetchUserData = (name: string | undefined) => {
-    const [userID, setUserID] = useState<number | null>(null);
-
-    useEffect(() => {
-        const fetchUserID = async () => {
-            try {
-                const response = await fetch(`${apiURL}/users?name=${name}`);
-                const data = await response.json();
-                const fetchedUserID = data.user?.id || data.id || null;
-                setUserID(fetchedUserID);
-            } catch (error) {
-                console.error("Error fetching userID:", error);
-            }
-        };
-
-        fetchUserID();
-    }, [name]);
-
-    return { userID };
-};
-
 export const fetchThreadCount = (
     isThread: boolean,
     userID?: number | null,
@@ -32,7 +11,6 @@ export const fetchThreadCount = (
 ) => {
     const token = localStorage.getItem("access-token");
     const { name } = useParams();
-    // const navigate = useNavigate();
     const [numOfThreads, setNumOfThreads] = useState<number>(0);
 
     const updateThreadCount = async () => {
@@ -58,10 +36,6 @@ export const fetchThreadCount = (
                 },
             );
             const data = await response.json();
-            // if (data.error == "Invalid user authentication") {
-            //     alert("User authentication failed. Please sign in again");
-            //     navigate("/");
-            // }
             setNumOfThreads(isThread ? data.total_threads : data.total_posts);
         } catch (error) {
             console.error("Error fetching thread count:", error);
@@ -77,20 +51,25 @@ export const fetchThreadCount = (
 
 export const handleDeleteClick = async (ID: number, isThread: boolean, name: string | undefined) => {
     const token = localStorage.getItem("access-token");
-    try {
-        const response = await fetch(`${apiURL}/${isThread ? "threads" : "posts"}/${ID}`, {
-            method: "DELETE",
-            headers: {
-                "Content-Type": "application/json",
-                Name: name,
-                Authorization: `Bearer ${token} `,
-            } as HeadersInit,
-        });
+    if (confirm("Are you want to delete?")) {
+        //Confirms decision to delete
+        try {
+            const response = await fetch(`${apiURL}/${isThread ? "threads" : "posts"}/${ID}`, {
+                method: "DELETE",
+                headers: {
+                    "Content-Type": "application/json",
+                    Name: name,
+                    Authorization: `Bearer ${token} `,
+                } as HeadersInit,
+            });
 
-        const data = await response.json();
-        console.log("Response from server:", data);
-    } catch (error) {
-        console.error("Error:", error);
+            const data = await response.json();
+            console.log("Response from server:", data);
+        } catch (error) {
+            console.error("Error:", error);
+        }
+    } else {
+        return;
     }
 };
 
@@ -115,3 +94,24 @@ export const handleEditClick = async (ID: number, textInput: string, isThread: b
         console.error("Error:", error);
     }
 };
+
+//  export const fetchUserData = (name: string | undefined) => {
+//     const [userID, setUserID] = useState<number | null>(null);
+
+//     useEffect(() => {
+//         const fetchUserID = async () => {
+//             try {
+//                 const response = await fetch(`${apiURL}/users?name=${name}`);
+//                 const data = await response.json();
+//                 const fetchedUserID = data.user?.id || data.id || null;
+//                 setUserID(fetchedUserID);
+//             } catch (error) {
+//                 console.error("Error fetching userID:", error);
+//             }
+//         };
+
+//         fetchUserID();
+//     }, [name]);
+
+//     return { userID };
+// };
